PixelPaint: Move GUI logic and filter parameters out of filters
