+++ Applications/PixelPaint/Filters/SpatialGaussianBlurFilter.cpp
@@ -38,26 +38,6 @@ SpatialGaussianBlurFilter<N, T>::~SpatialGaussianBlurFilter()
 {
 }
 
-template<size_t N, typename _T>
-OwnPtr<typename GenericConvolutionFilter<N>::Parameters>
-SpatialGaussianBlurFilter<N, _T>::get_parameters(Gfx::Bitmap& bitmap, const Gfx::IntRect& rect)
-{
-    Matrix<N, float> kernel;
-    auto sigma = 1.0f;
-    auto s = 2.0f * sigma * sigma;
-
-    for (auto x = -(ssize_t)N / 2; x <= (ssize_t)N / 2; x++) {
-        for (auto y = -(ssize_t)N / 2; y <= (ssize_t)N / 2; y++) {
-            auto r = sqrt(x * x + y * y);
-            kernel.elements()[x + 2][y + 2] = (exp(-(r * r) / s)) / (M_PI * s);
-        }
-    }
-
-    normalize(kernel);
-
-    return make<typename GenericConvolutionFilter<N>::Parameters>(bitmap, rect, kernel);
-}
-
 }
 
 template class PixelPaint::SpatialGaussianBlurFilter<3>;
