+++ Applications/PixelPaint/main.cpp
@@ -25,11 +25,7 @@
  */
 
 #include "CreateNewLayerDialog.h"
-#include "Filters/BoxBlurFilter.h"
-#include "Filters/GenericConvolutionFilter.h"
-#include "Filters/LaplacianFilter.h"
-#include "Filters/SharpenFilter.h"
-#include "Filters/SpatialGaussianBlurFilter.h"
+#include "FilterParams.h"
 #include "Image.h"
 #include "ImageEditor.h"
 #include "Layer.h"
@@ -223,14 +219,14 @@ int main(int argc, char** argv)
     edge_detect_submenu.add_action(GUI::Action::create("Laplacian (cardinal)", [&](auto&) {
         if (auto* layer = image_editor.active_layer()) {
             PixelPaint::LaplacianFilter filter;
-            if (auto parameters = filter.get_parameters(layer->bitmap(), layer->rect(), false))
+            if (auto parameters = PixelPaint::FilterParameters<PixelPaint::LaplacianFilter>::get(layer->bitmap(), layer->rect(), false))
                 filter.apply(*parameters);
         }
     }));
     edge_detect_submenu.add_action(GUI::Action::create("Laplacian (diagonal)", [&](auto&) {
         if (auto* layer = image_editor.active_layer()) {
             PixelPaint::LaplacianFilter filter;
-            if (auto parameters = filter.get_parameters(layer->bitmap(), layer->rect(), true))
+            if (auto parameters = PixelPaint::FilterParameters<PixelPaint::LaplacianFilter>::get(layer->bitmap(), layer->rect(), true))
                 filter.apply(*parameters);
         }
     }));
@@ -238,35 +234,35 @@ int main(int argc, char** argv)
     blur_submenu.add_action(GUI::Action::create("Gaussian Blur (3x3)", [&](auto&) {
         if (auto* layer = image_editor.active_layer()) {
             PixelPaint::SpatialGaussianBlurFilter<3> filter;
-            if (auto parameters = filter.get_parameters(layer->bitmap(), layer->rect()))
+            if (auto parameters = PixelPaint::FilterParameters<PixelPaint::SpatialGaussianBlurFilter<3>>::get(layer->bitmap(), layer->rect()))
                 filter.apply(*parameters);
         }
     }));
     blur_submenu.add_action(GUI::Action::create("Gaussian Blur (5x5)", [&](auto&) {
         if (auto* layer = image_editor.active_layer()) {
             PixelPaint::SpatialGaussianBlurFilter<5> filter;
-            if (auto parameters = filter.get_parameters(layer->bitmap(), layer->rect()))
+            if (auto parameters = PixelPaint::FilterParameters<PixelPaint::SpatialGaussianBlurFilter<5>>::get(layer->bitmap(), layer->rect()))
                 filter.apply(*parameters);
         }
     }));
     blur_submenu.add_action(GUI::Action::create("Box Blur (3x3)", [&](auto&) {
         if (auto* layer = image_editor.active_layer()) {
             PixelPaint::BoxBlurFilter<3> filter;
-            if (auto parameters = filter.get_parameters(layer->bitmap(), layer->rect()))
+            if (auto parameters = PixelPaint::FilterParameters<PixelPaint::BoxBlurFilter<3>>::get(layer->bitmap(), layer->rect()))
                 filter.apply(*parameters);
         }
     }));
     blur_submenu.add_action(GUI::Action::create("Box Blur (5x5)", [&](auto&) {
         if (auto* layer = image_editor.active_layer()) {
             PixelPaint::BoxBlurFilter<5> filter;
-            if (auto parameters = filter.get_parameters(layer->bitmap(), layer->rect()))
+            if (auto parameters = PixelPaint::FilterParameters<PixelPaint::BoxBlurFilter<5>>::get(layer->bitmap(), layer->rect()))
                 filter.apply(*parameters);
         }
     }));
     blur_submenu.add_action(GUI::Action::create("Sharpen", [&](auto&) {
         if (auto* layer = image_editor.active_layer()) {
             PixelPaint::SharpenFilter filter;
-            if (auto parameters = filter.get_parameters(layer->bitmap(), layer->rect()))
+            if (auto parameters = PixelPaint::FilterParameters<PixelPaint::SharpenFilter>::get(layer->bitmap(), layer->rect()))
                 filter.apply(*parameters);
         }
     }));
@@ -275,7 +271,7 @@ int main(int argc, char** argv)
     spatial_filters_menu.add_action(GUI::Action::create("Generic 5x5 Convolution", [&](auto&) {
         if (auto* layer = image_editor.active_layer()) {
             PixelPaint::GenericConvolutionFilter<5> filter;
-            if (auto parameters = filter.get_parameters(layer->bitmap(), layer->rect(), window))
+            if (auto parameters = PixelPaint::FilterParameters<PixelPaint::GenericConvolutionFilter<5>>::get(layer->bitmap(), layer->rect(), window))
                 filter.apply(*parameters);
         }
     }));
